Refactor more by putting general function into extension target
