+++ Gems/ScriptCanvas/Code/Include/ScriptCanvas/Data/DataTrait.h
@@ -11,41 +11,12 @@
 #include <ScriptCanvas/Data/BehaviorContextObjectPtr.h>
 #include <ScriptCanvas/Data/Constants.h>
 #include <ScriptCanvas/Data/Data.h>
+#include <ScriptCanvas/Data/DataTraitBase.h>
 
 namespace ScriptCanvas
 {
     namespace Data
-    {   
-        template<typename t_Type>
-        struct TraitsBase
-        {
-            using ThisType = TraitsBase<t_Type>;
-            using Type = AZStd::decay_t<t_Type>;
-            static const bool s_isAutoBoxed = false;
-            static const bool s_isKey = false;
-            static const bool s_isNative = false;
-            static const eType s_type = eType::Invalid;
-
-            static AZ::Uuid GetAZType(const Data::Type& = {}) { return azrtti_typeid<t_Type>(); }
-            static Data::Type GetSCType(const AZ::TypeId& = AZ::TypeId::CreateNull()) { return Data::FromAZType(GetAZType()); }
-            static AZStd::string GetName(const Data::Type& = {}) { return Data::GetName(Data::FromAZType(GetAZType())); }
-            // The static_assert needs to rely on the template parameter in order to avoid the clang frontend from asserting when parsing the template declaration
-            static Type GetDefault(const Data::Type& = {}) { static_assert((!AZStd::is_same<t_Type, t_Type>::value), "implement in the typed function"); return {}; }
-            static bool IsDefault(const AZStd::any&, const Data::Type& = {}) { static_assert((!AZStd::is_same<t_Type, t_Type>::value), "implement in the typed function"); return {}; }
-        };
-
-        
-        template<typename t_Type>
-        struct Traits : public TraitsBase<t_Type>
-        {
-        };
-                
-        // a compile time map of eType back to underlying AZ type and traits
-        template<eType>
-        struct eTraits
-        {
-        };
-
+    {                   
         struct TypeErasedDataTraits
         {
             AZ_CLASS_ALLOCATOR(TypeErasedDataTraits, AZ::SystemAllocator, 0);
