+++ Gems/ScriptCanvas/Code/Include/ScriptCanvas/Libraries/Core/ContainerTypeReflection.h
@@ -8,6 +8,7 @@
 
 #include <AzCore/Script/ScriptContextAttributes.h>
 #include <ScriptCanvas/Data/DataMacros.h>
+#include <ScriptCanvas/Data/DataTraitBase.h>
 
 #pragma once
